fixed release and API doc generation, set release version and date for fleaTLS version 1.2

+++ misc/tools/generate_release.py
@@ -126,9 +126,10 @@ have_test_data = True
 #        sys.exit(1)
 
 
-shutil.rmtree(generate_dir + "/" + license_name_gpl + "/" + "flea", True)
-shutil.rmtree(generate_dir + "/" + license_name_closed_source + "/" + "flea", True)
-shutil.rmtree(generate_dir)
+#shutil.rmtree(generate_dir + "/" + license_name_gpl + "/" + "flea", True)
+#shutil.rmtree(generate_dir + "/" + license_name_closed_source + "/" + "flea", True)
+if os.path.exists(generate_dir):
+  shutil.rmtree(generate_dir)
 
 generate_with_license(license_name_gpl, have_test_data)
 generate_with_license(license_name_closed_source, have_test_data)
